AI_Job_Hunter/requirements.txt
streamlit
requests
python-dotenv
PyPDF2
pdfminer.six
ollama
